make ToggleGroup composable
