+++ packages/app/src/lib/components/ToggleGroup.svelte
@@ -1,7 +1,31 @@
+<!--
+	ToggleGroup - A composable wrapper around Melt UI's toggle group
+
+	This component handles the complexity of integrating Melt UI's toggle group with 
+	custom rendering via Svelte 5 snippets. The architecture requires:
+
+	1. DATA: Pass items array to initialize Melt UI state management
+	2. RENDERING: Use children snippet to customize how each item renders
+	3. CONSTRAINTS: Melt UI requires use:melt on HTML buttons (not components)
+	
+	The apparent "redundancy" of passing items AND rendering them via snippet is
+	necessary because:
+	- Melt UI needs the complete data array for state/accessibility management
+	- We need rendering flexibility without losing toggle functionality
+	- Can't use pure composition due to Svelte's use: directive limitations
+
+	Usage:
+	<ToggleGroup items={tags} selectedItems={selected}>
+		{#snippet children(item, isSelected, isDisabled)}
+			<Tag variant={isSelected ? 'selected' : 'default'}>{item}</Tag>
+		{/snippet}
+	</ToggleGroup>
+-->
 <script lang="ts">
 	import { createToggleGroup, melt } from '@melt-ui/svelte';
 	import { Check } from 'phosphor-svelte';
 	import { mergeCss } from '$utils/utils';
+	import type { Snippet } from 'svelte';
 
 	interface Props {
 		items: string[];
@@ -11,6 +35,7 @@
 		disabledItems?: string[];
 		onItemSelected?: (selected: string[]) => void;
 		class?: string;
+		children: Snippet<[item: string, isSelected: boolean, isDisabled: boolean]>;
 	}
 
 	let {
@@ -20,7 +45,8 @@
 		orientation = 'vertical',
 		type = 'multiple',
 		onItemSelected,
-		class: className
+		class: className,
+		children
 	}: Props = $props();
 
 	let sortedItems = $derived([...items].sort());
@@ -39,6 +65,10 @@
 			return next;
 		}
 	});
+
+	function isSelected(itemValue: string): boolean {
+		return $value && ($value.includes(itemValue) || (type === 'single' && $value === itemValue));
+	}
 
 	function isDisabled(itemValue: string): boolean {
 		return disabledItems.includes(itemValue);
@@ -60,9 +90,9 @@
 			use:melt={$item({ value: itemValue, disabled: isDisabled(itemValue) })}
 			aria-label="Toggle {itemValue}"
 			role={type === 'single' ? 'radio' : 'checkbox'}
-			aria-checked={($value && $value.includes(itemValue)) ||
-				(type === 'single' && $value === itemValue)}
+			aria-checked={isSelected(itemValue)}
 		>
+			<!-- Checkbox (still generated by ToggleGroup) -->
 			<span
 				class="w-5 h-5 border-2 border-gray-300 rounded flex items-center justify-center bg-white text-blue-600 transition-colors data-[state='on']:border-blue-600 data-[state='on']:bg-blue-600 data-[state='on']:text-white"
 				class:border-gray-200={isDisabled(itemValue)}
@@ -70,17 +100,13 @@
 				class:text-gray-400={isDisabled(itemValue)}
 				aria-hidden="true"
 			>
-				{#if ($value && $value.includes(itemValue)) || (type === 'single' && $value === itemValue)}
-					<Check size={32} weight="bold" />
+				{#if isSelected(itemValue)}
+					<Check size={16} weight="bold" />
 				{/if}
 			</span>
-			<span
-				class="text-sm text-gray-700 transition-colors select-none"
-				class:text-gray-400={isDisabled(itemValue)}
-				aria-hidden="true"
-			>
-				{itemValue}
-			</span>
+
+			<!-- Composable content area -->
+			{@render children(itemValue, isSelected(itemValue), isDisabled(itemValue))}
 		</button>
 	{/each}
 </div>
