--- packages/app/src/lib/components/ToggleGroupComposable.svelte
@@ -1,112 +1,0 @@
-<!--
-	ToggleGroupComposable - A composable wrapper around Melt UI's toggle group
-
-	This component handles the complexity of integrating Melt UI's toggle group with 
-	custom rendering via Svelte 5 snippets. The architecture requires:
-
-	1. DATA: Pass items array to initialize Melt UI state management
-	2. RENDERING: Use children snippet to customize how each item renders
-	3. CONSTRAINTS: Melt UI requires use:melt on HTML buttons (not components)
-	
-	The apparent "redundancy" of passing items AND rendering them via snippet is
-	necessary because:
-	- Melt UI needs the complete data array for state/accessibility management
-	- We need rendering flexibility without losing toggle functionality
-	- Can't use pure composition due to Svelte's use: directive limitations
-
-	Usage:
-	<ToggleGroupComposable items={tags} selectedItems={selected}>
-		{#snippet children(item, isSelected, isDisabled)}
-			<Tag variant={isSelected ? 'selected' : 'default'}>{item}</Tag>
-		{/snippet}
-	</ToggleGroupComposable>
--->
-<script lang="ts">
-	import { createToggleGroup, melt } from '@melt-ui/svelte';
-	import { Check } from 'phosphor-svelte';
-	import { mergeCss } from '$utils/utils';
-	import type { Snippet } from 'svelte';
-
-	interface Props {
-		items: string[];
-		orientation?: 'horizontal' | 'vertical';
-		type?: 'single' | 'multiple';
-		selectedItems?: string[];
-		disabledItems?: string[];
-		onItemSelected?: (selected: string[]) => void;
-		class?: string;
-		children: Snippet<[item: string, isSelected: boolean, isDisabled: boolean]>;
-	}
-
-	let {
-		items,
-		selectedItems = [],
-		disabledItems = [],
-		orientation = 'vertical',
-		type = 'multiple',
-		onItemSelected,
-		class: className,
-		children
-	}: Props = $props();
-
-	let sortedItems = $derived([...items].sort());
-
-	const {
-		elements: { root, item },
-		states: { value }
-	} = createToggleGroup({
-		type,
-		defaultValue: type === 'single' ? selectedItems[0] : selectedItems,
-		orientation: orientation,
-		onValueChange: ({ curr, next }) => {
-			if (onItemSelected) {
-				onItemSelected(next);
-			}
-			return next;
-		}
-	});
-
-	function isSelected(itemValue: string): boolean {
-		return $value && ($value.includes(itemValue) || (type === 'single' && $value === itemValue));
-	}
-
-	function isDisabled(itemValue: string): boolean {
-		return disabledItems.includes(itemValue);
-	}
-</script>
-
-<div
-	use:melt={$root}
-	class={mergeCss(
-		"flex items-start data-[orientation='vertical']:flex-col data-[orientation='horizontal']:flex-row gap-1",
-		className
-	)}
-	aria-label="Toggle selection"
-	role={type === 'single' ? 'radiogroup' : 'group'}
->
-	{#each sortedItems as itemValue (itemValue)}
-		<button
-			class="pb-1 last-child:pb-0 flex items-center gap-2 w-full text-left cursor-pointer transition-colors hover:bg-gray-50 focus:outline-none focus:bg-gray-50 data-[disabled]:cursor-not-allowed data-[disabled]:opacity-50 data-[disabled]:hover:bg-transparent"
-			use:melt={$item({ value: itemValue, disabled: isDisabled(itemValue) })}
-			aria-label="Toggle {itemValue}"
-			role={type === 'single' ? 'radio' : 'checkbox'}
-			aria-checked={isSelected(itemValue)}
-		>
-			<!-- Checkbox (still generated by ToggleGroup) -->
-			<span
-				class="w-5 h-5 border-2 border-gray-300 rounded flex items-center justify-center bg-white text-blue-600 transition-colors data-[state='on']:border-blue-600 data-[state='on']:bg-blue-600 data-[state='on']:text-white"
-				class:border-gray-200={isDisabled(itemValue)}
-				class:bg-gray-100={isDisabled(itemValue)}
-				class:text-gray-400={isDisabled(itemValue)}
-				aria-hidden="true"
-			>
-				{#if isSelected(itemValue)}
-					<Check size={16} weight="bold" />
-				{/if}
-			</span>
-
-			<!-- Composable content area -->
-			{@render children(itemValue, isSelected(itemValue), isDisabled(itemValue))}
-		</button>
-	{/each}
-</div>
